requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
backports.zoneinfo==0.2.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
cloudinary==1.32.0
Django==4.1.7
djangorestframework==3.14.0
flatbuffers==23.1.21
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
imageio==2.25.1
importlib-metadata==6.0.0
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.10.1
six==1.16.0
scikit-learn
sqlparse==0.4.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tifffile==2023.2.3
typing-extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.14.0
